fix(sdcard.device): report TDERR_BadUnitNum when unit not found
The intention according to the comment is not to continue if the unit is
not found, so we should set the error code and return immediately.

This is the only error case in function, so the conditionals around the
rest of the process become redundant.

That TDERR_BadUnitNum rather than IOERR_OPENFAIL is the correct code for
the error case was determined in conversation with _Bnu on the PiStorm
Discord.